Bug fix attempt 2

+++ bomb.h
@@ -31,10 +31,10 @@ typedef struct BombSettings {
     bool switchRedBomb;
     bool spawnInside;
     int explosionRange;
-    uint8_t boxIdxesLeft[2];
-    uint8_t boxIdxesRight[2];
-    uint8_t boxIdxesTop[2];
-    uint8_t boxIdxesBottom[2];
+    int boxIdxesLeft[2];
+    int boxIdxesRight[2];
+    int boxIdxesTop[2];
+    int boxIdxesBottom[2];
 } *Bomb;
 
 void bombPlacement(Player p, Bomb bombs[], SDL_Renderer *renderer);
